Add LST restaking to Solayer
